move retore method from CatalogObjectController to CatalogObjectRevisionController
move retore method from CatalogObjectController to CatalogObjectRevisionController

move retore method from CatalogObjectController to CatalogObjectRevisionController

move restore method from CatalogObjectController to CatalogObjectRevisionController and refactor code
